Remove `PassphraseDialogManager` and move logic to `RestoreLocalBackupFragment`
